pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "poolmind"
dynamic = ["version"]
description = "AI-powered Cross-Exchange Arbitrage Trading System"
readme = "README.md"
authors = [
    {name = "PoolMind Team", email = "info@poolmind.ai"},
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business :: Financial :: Investment",
]
keywords = ["crypto", "trading", "arbitrage", "ai", "llm", "multi-agent"]
dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "asyncio>=3.4.3",
    "groq>=0.4.0",
    "langchain>=0.1.0",
    "langgraph>=0.0.20",
    "chromadb>=0.4.18",
    "ccxt>=4.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "redis>=4.6.0",
    "psycopg2-binary>=2.9.6",
    "loguru>=0.7.0",
    "tenacity>=8.2.3",
    "httpx>=0.24.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.1",
    "flake8>=6.1.0",
]
docs = [
    "sphinx>=7.1.2",
    "sphinx-rtd-theme>=1.3.0",
]

[project.urls]
"Homepage" = "https://github.com/poolmind/poolmind"
"Bug Tracker" = "https://github.com/poolmind/poolmind/issues"
"Documentation" = "https://poolmind.readthedocs.io/"

[project.scripts]
poolmind = "poolmind.app:main"

[tool.setuptools]
packages = ["poolmind"]

[tool.setuptools.dynamic]
version = {attr = "poolmind.__version__"}

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
asyncio_mode = "auto"
